reset ticks type + structure of book class

+++ include/exchange-disruptor/matching_engine/event/fill_event.h
@@ -5,14 +5,14 @@
 
 class FillEvent : public Event {
 public:
-    FillEvent(uint64_t clientId, uint64_t orderId, uint64_t matchId, double price, double qty) : Event(clientId, orderId), _matchId(matchId), _price(price), _qty(qty) {}
+    FillEvent(uint64_t clientId, uint64_t orderId, uint64_t matchId, int64_t price, double qty) : Event(clientId, orderId), _matchId(matchId), _price(price), _qty(qty) {}
     EventType getType() const override;
     uint64_t getMatchId() const;
-    double getPrice() const;
+    int64_t getPrice() const;
     double getQty() const;
 private:
     uint64_t _matchId;
-    double _price;
+    int64_t _price;
     double _qty;
 };
